# api/doc/api-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Crée un nouvel utilisateur",
        "description": "A utiliser en 1er pour pouvoir se logger ensuite avec `/auth/login`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "password"
                ],
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur crée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Connecte un nouvel utilisateur",
        "description": "Permet de récupérer un token dans les cookies coté client. Il est utilisé dans toutes les routes  qui necessitent une authentification comme `/auth/me` ou `/auth/logout`.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utilisateur connecté",
            "headers": {
              "Set-Cookie": {
                "description": "Cookie contenant le token JWT nommé \"token\"",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Déconnecte l'utilisateur",
        "description": "Le client doit envoyer le token généré à partir de la route /auth/login en tant que cookie",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur connecté",
            "headers": {
              "Set-Cookie": {
                "description": "Cookie contenant le token JWT",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Renvoie les informations d'un utilisateur connecté",
        "description": "Le client doit envoyer le token généré à partir de la route /auth/login en tant que cookie",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "summary": "Renvoie la liste des catégories de l'utilisateur connecté",
        "description": "Seul les utilisateurs avec un token en cookie sont autorisés sur cette route. Seul les catégories de l'utilisateur connecté sont renvoyées.\n",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number"
                    },
                    "categories": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crée une nouvelle catégorie",
        "description": "Seul les utilisateurs avec un token en cookie sont autorisés sur cette route.\n",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nom de la catégorie (entre 2 et 50 caractères)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Crée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "summary": "Renvoie les informations sur une catégorie",
        "description": "Seul les utilisateurs avec un token en cookie sont autorisés sur cette route. Seulement les catégories appartenant à l'utilisateur peuvent être renvoyées.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identifiant de la catégorie",
            "schéma": {
              "type": "number"
            }
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Met à jour le nom de la catégorie",
        "description": "Seul les utilisateurs avec un token en cookie sont autorisés sur cette route. Seulement les catégories appartenant à l'utilisateur peuvent être modifiées.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identifiant de la catégorie",
            "schéma": {
              "type": "number"
            }
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nom de la catégorie (entre 2 et 50 caractères)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mise à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Supprime la catégorie",
        "description": "Seul les utilisateurs avec un token en cookie sont autorisés sur cette route. Seulement les catégories appartenant à l'utilisateur peuvent être supprimées.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identifiant de la catégorie",
            "schéma": {
              "type": "number"
            }
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Supprimée avec succès"
          }
        }
      }
    },
    "/transactions": {
      "get": {
        "summary": "Renvoie la liste des transactions de l'utilisateur connecté",
        "description": "Seul les utilisateurs avec un token en cookie sont autorises sur cette route. Seul les transactions de l'utilisateur connecté sont renvoyées.\n",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number"
                    },
                    "transactions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "label": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "date": {
                            "type": "string"
                          },
                          "categoryId": {
                            "type": "number"
                          },
                          "userId": {
                            "type": "number"
                          },
                          "description": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cree une nouvelle transaction",
        "description": "Seul les utilisateurs avec un token en cookie sont autorisés sur cette route.\n",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "label",
                  "date",
                  "categoryId"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "Montant de la transaction (strictement positif)."
                  },
                  "label": {
                    "type": "string",
                    "description": "Libelle de la transaction."
                  },
                  "date": {
                    "type": "string",
                    "description": "Date au format ISO datetime."
                  },
                  "categoryId": {
                    "type": "number",
                    "description": "Identifiant de la categorie (entier positif)."
                  },
                  "description": {
                    "type": "string",
                    "description": "Informations supplémentaires"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Crée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "label": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number"
                    },
                    "date": {
                      "type": "string"
                    },
                    "categoryId": {
                      "type": "number"
                    },
                    "userId": {
                      "type": "number"
                    },
                    "description": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{id}": {
      "get": {
        "summary": "Renvoie les informations sur une transaction",
        "description": "Seul les utilisateurs avec un token en cookie sont autorises sur cette route. Seulement les transactions appartenant a l'utilisateur peuvent etre renvoyees.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identifiant de la transaction",
            "schema": {
              "type": "number"
            }
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "label": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number"
                    },
                    "date": {
                      "type": "string"
                    },
                    "categoryId": {
                      "type": "number"
                    },
                    "userId": {
                      "type": "number"
                    },
                    "description": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Met a jour la transaction",
        "description": "Seul les utilisateurs avec un token en cookie sont autorises sur cette route. Seulement les transactions appartenant a l'utilisateur peuvent etre modifiees.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identifiant de la transaction",
            "schema": {
              "type": "number"
            }
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "Montant de la transaction (strictement positif)."
                  },
                  "label": {
                    "type": "string",
                    "description": "Libelle de la transaction."
                  },
                  "date": {
                    "type": "string",
                    "description": "Date au format ISO datetime."
                  },
                  "categoryId": {
                    "type": "number",
                    "description": "Identifiant de la categorie (entier positif)."
                  },
                  "description": {
                    "type": "string",
                    "description": "Informations supplémentaires."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mise a jour avec succes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Supprime la transaction",
        "description": "Seul les utilisateurs avec un token en cookie sont autorises sur cette route. Seulement les transactions appartenant a l'utilisateur peuvent etre supprimees.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identifiant de la transaction",
            "schema": {
              "type": "number"
            }
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Supprimée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}